+++ src/run.test.ts
@@ -2,7 +2,8 @@ import * as run from './run'
 import {
    getkubectlDownloadURL,
    getKubectlArch,
-   getExecutableExtension
+   getExecutableExtension,
+   getLatestPatchVersion
 } from './helpers'
 import * as os from 'os'
 import * as toolCache from '@actions/tool-cache'
@@ -11,7 +12,11 @@ import * as path from 'path'
 import * as core from '@actions/core'
 import * as util from 'util'
 
+
 describe('Testing all functions in run file.', () => {
+   beforeEach(() => {
+      jest.clearAllMocks()
+   })
    test('getExecutableExtension() - return .exe when os is Windows', () => {
       jest.spyOn(os, 'type').mockReturnValue('Windows_NT')
       expect(getExecutableExtension()).toBe('.exe')
@@ -164,28 +169,27 @@ describe('Testing all functions in run file.', () => {
       )
       expect(toolCache.downloadTool).not.toHaveBeenCalled()
    })
-
    test('getLatestPatchVersion() - download and return latest patch version', async () => {
       jest
          .spyOn(toolCache, 'downloadTool')
-         .mockReturnValue(Promise.resolve('pathToTool'))
+         .mockResolvedValue('pathToTool')
       jest.spyOn(fs, 'readFileSync').mockReturnValue('v1.27.15')
 
-      const result = await run.getLatestPatchVersion('1', '27')
+      const result = await getLatestPatchVersion('1', '27')
 
       expect(result).toBe('v1.27.15')
       expect(toolCache.downloadTool).toHaveBeenCalledWith(
          'https://cdn.dl.k8s.io/release/stable-1.27.txt'
       )
-      expect(fs.readFileSync).toHaveBeenCalledWith('pathToTool', 'utf8')
    })
+
    test('getLatestPatchVersion() - throw error when patch version is empty', async () => {
       jest
          .spyOn(toolCache, 'downloadTool')
-         .mockReturnValue(Promise.resolve('pathToTool'))
+         .mockResolvedValue('pathToTool')
       jest.spyOn(fs, 'readFileSync').mockReturnValue('')
 
-      await expect(run.getLatestPatchVersion('1', '27')).rejects.toThrow(
+      await expect(getLatestPatchVersion('1', '27')).rejects.toThrow(
          'Failed to get latest patch version for 1.27'
       )
    })
@@ -195,35 +199,36 @@ describe('Testing all functions in run file.', () => {
          .spyOn(toolCache, 'downloadTool')
          .mockRejectedValue(new Error('Network error'))
 
-      await expect(run.getLatestPatchVersion('1', '27')).rejects.toThrow(
+      await expect(getLatestPatchVersion('1', '27')).rejects.toThrow(
          'Failed to get latest patch version for 1.27'
       )
    })
    test('resolveKubectlVersion() - expands major.minor to latest patch', async () => {
-      // Mock the getLatestPatchVersion call
-      jest.spyOn(run, 'getLatestPatchVersion').mockResolvedValue('v1.27.15')
+      jest
+         .spyOn(toolCache, 'downloadTool')
+         .mockResolvedValue('pathToTool')
+      jest.spyOn(fs, 'readFileSync').mockReturnValue('v1.27.15')
 
       const result = await run.resolveKubectlVersion('1.27')
-
       expect(result).toBe('v1.27.15')
-      expect(run.getLatestPatchVersion).toHaveBeenCalledWith('1', '27')
    })
+
    test('resolveKubectlVersion() - returns full version unchanged', async () => {
       const result = await run.resolveKubectlVersion('v1.27.15')
       expect(result).toBe('v1.27.15')
    })
-
    test('resolveKubectlVersion() - adds v prefix to full version', async () => {
       const result = await run.resolveKubectlVersion('1.27.15')
       expect(result).toBe('v1.27.15')
    })
    test('resolveKubectlVersion() - expands v-prefixed major.minor to latest patch', async () => {
-      jest.spyOn(run, 'getLatestPatchVersion').mockResolvedValue('v1.27.15')
+      jest
+         .spyOn(toolCache, 'downloadTool')
+         .mockResolvedValue('pathToTool')
+      jest.spyOn(fs, 'readFileSync').mockReturnValue('v1.27.15')
 
       const result = await run.resolveKubectlVersion('v1.27')
-
       expect(result).toBe('v1.27.15')
-      expect(run.getLatestPatchVersion).toHaveBeenCalledWith('1', '27')
    })
    test('run() - download specified version and set output', async () => {
       jest.spyOn(core, 'getInput').mockReturnValue('v1.15.5')
